"Applied fix from trunk framework for revision: 1832662  " ------------------------------------------------------------------------ r1832662 | jleroux | 2018-06-01 09:37:27 +0200 (ven. 01 juin 2018) | 12 lignes
Fixes a session fixation security issue discovered by a client with the security
 audit tool "IBM Security AppScan Enterprise , Version : 9.0.3.7"

Prevents the session fixation by making Tomcat generate a new jsessionId 
(ultimately put in cookie). 

Only do when really signing in to avoid unnecessary calls
Though if the client has disabled the use of cookies, then a session will be 
new on each request, not a good choice on client side!



------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release16.11@1832664 13f79535-47bb-0310-9956-ffa450edef68